+++ firebase-vscode/src/data-connect/index.ts
@@ -28,6 +28,7 @@ import { runDataConnectCompiler } from "./core-compiler";
 import { Result } from "../result";
 import { runEmulatorIssuesStream } from "./emulator-stream";
 import { LanguageClient } from "vscode-languageclient/node";
+import { registerTerminalTasks } from "./terminal";
 
 class CodeActionsProvider implements vscode.CodeActionProvider {
   constructor(
@@ -153,17 +154,19 @@ export function registerFdc(
   );
   const schemaCodeLensProvider = new SchemaCodeLensProvider(emulatorController);
 
+  // activate language client/serer
   let client: LanguageClient;
+  const lsOutputChannel: vscode.OutputChannel = vscode.window.createOutputChannel(
+    "Firebase GraphQL Language Server",
+  );
+
   // setup new language client on config change
   context.subscriptions.push({
     dispose: effect(() => {
       const configs = dataConnectConfigs.value?.tryReadValue;
       if (client) client.stop();
       if (configs && configs.values.length > 0) {
-        client = setupLanguageClient(
-          context,
-          configs,
-        );
+        client = setupLanguageClient(context, configs, lsOutputChannel);
         vscode.commands.executeCommand("fdc-graphql.start");
       }
     }),
@@ -179,7 +182,7 @@ export function registerFdc(
         vscode.commands.executeCommand(
           "firebase.dataConnect.executeIntrospection",
         );
-        runEmulatorIssuesStream(configs,fdcService.localEndpoint.value);
+        runEmulatorIssuesStream(configs, fdcService.localEndpoint.value);
         runDataConnectCompiler(configs, fdcService.localEndpoint.value);
       }
     }),
@@ -205,7 +208,7 @@ export function registerFdc(
   return Disposable.from(
     codeActions,
     selectedProjectStatus,
-    {dispose: sub1},
+    { dispose: sub1 },
     {
       dispose: effect(() => {
         selectedProjectStatus.text = `$(mono-firebase) ${
@@ -214,13 +217,14 @@ export function registerFdc(
         selectedProjectStatus.show();
       }),
     },
-    registerDataConnectConfigs(),
+    registerDataConnectConfigs(broker),
     registerExecution(context, broker, fdcService, emulatorController),
     registerExplorer(context, broker, fdcService),
     registerFirebaseDataConnectView(context, broker, emulatorController),
-    registerAdHoc(),
+    registerAdHoc(fdcService),
     registerConnectors(context, broker, fdcService),
     registerFdcDeploy(broker),
+    registerTerminalTasks(broker),
     operationCodeLensProvider,
     vscode.languages.registerCodeLensProvider(
       // **Hack**: For testing purposes, enable code lenses on all graphql files